- replace create_or_empty_folder function pathlib.Path functions - move reader functions in utils.py to readers.py in utils folder - (previous commit) add verifications.py with, for ex., functions to check validity of features in vector files - add lots of progress bars with tqdm in for loop all around images_to_samples.py and utils.py
