+++ utils/verifications.py
@@ -1,3 +1,9 @@
+import fiona
+from tqdm import tqdm
+
+from utils.utils import get_key_recursive
+
+
 def is_valid_geom(geom):
     """
     Checks to see if geometry is a valid GeoJSON geometry type or
@@ -87,3 +93,28 @@ def is_valid_geom(geom):
                 return False  # short-circuit and fail early
 
     return True
+
+
+def validate_num_classes(vector_file, num_classes, attribute_name, ignore_index, debug=True):    # used only in images_to_samples.py
+    """Validate that the number of classes in the vector file corresponds to the expected number
+    Args:
+        vector_file: full file path of the vector image
+        num_classes: number of classes set in config.yaml
+        attribute_name: name of the value field representing the required classes in the vector image file
+
+        value_field: name of the value field representing the required classes in the vector image file
+        ignore_index: (int) target value that is ignored during training and does not contribute to the input gradient
+    Return:
+        None
+    """
+
+    distinct_att = set()
+    with fiona.open(vector_file, 'r') as src:
+        for feature in tqdm(src, leave=False, position=1, desc=f'Scanning features'):
+            distinct_att.add(get_key_recursive(attribute_name, feature))  # Use property of set to store unique values
+
+    detected_classes = len(distinct_att) - len([ignore_index]) if ignore_index in distinct_att else len(distinct_att)
+
+    if detected_classes != num_classes:
+        raise ValueError('The number of classes in the yaml.config {} is different than the number of classes in '
+                         'the file {} {}'.format (num_classes, vector_file, str(list(distinct_att))))
